pick files

+++ app/src/main/java/pro/aidar/library/data/dto/Book.kt
@@ -9,7 +9,7 @@ data class Book(
     @PrimaryKey(autoGenerate = true)
     var id: Int? = null,
     var name: String? = null,
-    var size: Int? = null,
+    var size: Long? = null,
     var updateDate: Date? = null,
     var author: String? = null,
     var genre: String? = null,
